// src/data/types.ts
import { Route } from "../routers/types";
import { StaticImageData } from "next/image";

//  ######  CustomLink  ######## //
export interface CustomLink {
  label: string;
  href: Route;
  targetBlank?: boolean;
}

//  ##########  PostDataType ######## //
export interface TaxonomyType {
  id: string | number;
  name: string;
  href: Route;
  count?: number;
  thumbnail?: string | StaticImageData;
  desc?: string;
  color?: TwMainColor | string;
  taxonomy: "category" | "tag";
}

export interface PostAuthorType {
  id?: string | number;
  firstName?: string;
  lastName?: string;
  displayName: string;
  avatar: string | StaticImageData;
  bgImage?: string | StaticImageData;
  email?: string;
  count?: number;
  desc?: string;
  jobName?: string;
  href?: string;
}

export interface PostDataType {
  id: string | number;
  author?: PostAuthorType;
  date: string;
  href: string;
  categories?: TaxonomyType[];
  title: string;
  created_at:any
  cover_photo_url:String
  video_link:String
  featuredImage?: string | StaticImageData;
  desc?: string;
  companyName?: string; 
  member?: string;
  like: {
    count: number;
    isLiked: boolean;
  };
  bookmark?: {
    count: number;
    isBookmarked: boolean;
  };
  commentCount?: number;
  viewdCount?: number;
  readingTime?: number;
  postType?: "standard" | "video" | "gallery" | "audio";
  videoUrl?: string;
  audioUrl?: string | string[];
  galleryImgs?: string[];
}

export type TwMainColor =
  | "pink"
  | "green"
  | "yellow"
  | "red"
  | "indigo"
  | "blue"
  | "purple"
  | "gray";